Disable ssh connection multiplexing
Connection multiplexing can result in the program hanging waiting for
a newline on stdin if something is typed on the console when this is
run.  Disabling multiplexing seems to reliably prevent this bug from
happening.

Reproduction: `snapsend config.sexp` quickly using Up Arrow, Enter.
If the up arrow key is pressed at the wrong moment while an ssh
process is running, `snapsend` hangs indefinitely.